+++ src/main/scala/chisel3/util/experimental/minimizer/QMCMinimizer.scala
@@ -4,102 +4,11 @@ import chisel3.util.BitPat
 
 import scala.annotation.tailrec
 import scala.math.Ordered.orderingToOrdered
+import Minimizer._
 
 object QMCMinimizer {
   /** @return A instance of [[QMCMinimizer]] */
   def apply(): QMCMinimizer = new QMCMinimizer()
-
-  implicit class Implicant(x: BitPat) {
-    /** Check whether two implicants have the same value on all of the cared bits (intersection).
-      *
-      * {{{
-      * value ^^ x.value                                       // bits that are different
-      * (bits that are different) & x.mask                     // bits that are different and `this` care
-      * (bits that are different and `this` care) & y.mask     // bits that are different and `both` care
-      * (bits that are different and both care) == 0           // no (bits that are different and we both care) exists
-      * no (bits that are different and we both care) exists   // all cared bits are the same, two terms intersect
-      * }}}
-      *
-      * @param y Implicant to be checked with
-      * @return Whether two implicants intersect
-      */
-    def intersects(y: BitPat): Boolean = ((x.value ^ y.value) & x.mask & y.mask) == 0
-
-    /** Check if two implicants are the same
-      * @param y Implicant to be checked with
-      * @return Whether two implicants are the same
-      */
-    def sameAs(y: BitPat): Boolean = x.mask == y.mask && x.value == y.value
-
-    /** Merge two similar implicants if they are similar
-      * Rule of merging: '0' and '1' merge to '?'
-      * Two implicants are "similar" when they satisfy all the following rules:
-      *   1. have the same mask ('?'s are at the same positions)
-      *   1. values only differ by one bit
-      *
-      * @example this = 11?0, x = 10?0 -> similar
-      * @example this = 11??, x = 10?0 -> not similar, violated rule 1
-      * @example this = 11?1, x = 10?0 -> not similar, violated rule 2
-      * @param y Implicant to be merged with
-      * @return Merge result wrapped in [[Some]] if `x` and `y` are similar,
-      *         [[None]] otherwise
-      */
-    def mergeIfSimilar(y: BitPat): Option[BitPat] = {
-      val diff = x.value ^ y.value
-      if (x.mask == y.mask && diff.bitCount == 1) { // similar
-        Some(new BitPat(x.value &~ diff, x.mask &~ diff, x.getWidth))
-      } else {
-        None
-      }
-    }
-
-    /** Check all bits in `x` cover the correspond position in `y`.
-      *
-      * Rule to define coverage relationship among `0`, `1` and `?`:
-      *   1. '?' covers '0' and '1', '0' covers '0', '1' covers '1'
-      *   1. '1' doesn't cover '?', '1' doesn't cover '0'
-      *   1. '0' doesn't cover '?', '0' doesn't cover '1'
-      *
-      * For all bits that `x` don't care, `y` can be `0`, `1`, `?`
-      * For all bits that `x` care, `y` must be the same value and not masked.
-      * {{{
-      *    (~x.mask & -1) | ((x.mask) & ((x.value xnor y.value) & y.mask)) = -1
-      * -> ~x.mask | ((x.mask) & ((x.value xnor y.value) & y.mask)) = -1
-      * -> ~x.mask | ((x.value xnor y.value) & y.mask) = -1
-      * -> x.mask & ~((x.value xnor y.value) & y.mask) = 0
-      * -> x.mask & (~(x.value xnor y.value) | ~y.mask) = 0
-      * -> x.mask & ((x.value ^ y.value) | ~y.mask) = 0
-      * -> ((x.value ^ y.value) & x.mask | ~y.mask & x.mask) = 0
-      * }}}
-      *
-      * @param y to check is covered by `x` or not.
-      * @return Whether `x` covers `y`
-      */
-    def covers(y: BitPat): Boolean = ((x.value ^ y.value) & x.mask | ~y.mask & x.mask) == 0
-
-    /** Expand implicant `x` to a [[Seq]] of implicants without touching the forbidden implicants specified by `maxt`.
-      * @param maxt The forbidden list
-      * @return     Expanded implicants
-      */
-    def expand(maxt: Seq[BitPat]): Seq[BitPat] = (0 until x.getWidth).flatMap{ i =>
-        val newImplicant = x match {
-          case x if x.mask.testBit(i) && !x.value.testBit(i) => // this bit is 0
-            Some(new BitPat(x.value.setBit(i), x.mask, x.getWidth))
-          case x if x.mask.testBit(i) && x.value.testBit(i) => // this bit is 1
-            Some(new BitPat(x.value.clearBit(i), x.mask, x.getWidth))
-          case x if !x.mask.testBit(i) => // this bit is ?
-            None
-        }
-        newImplicant.flatMap(a => if (maxt.exists(a.intersects(_))) None else Some(a))
-      }
-  }
-
-  /**
-    * If two terms have different value, then their order is determined by the value, or by the mask.
-    */
-  implicit def ordering: Ordering[BitPat] = (x: BitPat, y: BitPat) => {
-    if (x.value < y.value || x.value == y.value && x.mask > y.mask) -1 else 1
-  }
 
   /** Calculate essential prime implicants based on previously calculated prime implicants and all implicants.
     *
@@ -162,7 +71,7 @@ object QMCMinimizer {
       *
       * @param a    Operand a
       * @param b    Operand b
-      * @return
+      * @return TODO
       */
     def cheaper(a: Seq[BitPat], b: Seq[BitPat]): Boolean = {
       val ca = getCost(a)
@@ -199,9 +108,8 @@ object QMCMinimizer {
 /** The [[https://en.wikipedia.org/wiki/Quine-McCluskey_algorithm]] decoder implementation. */
 class QMCMinimizer extends Minimizer {
   def minimize(default: BitPat, table: Seq[(BitPat, BitPat)]): Seq[Seq[BitPat]] = {
-    import QMCMinimizer.Implicant
 
-    require(table.nonEmpty, "Truth table must not be empty")
+    require(table.nonEmpty, "Truth table must not be empty") // TODO return `default`
 
     // extract decode table to inputs and outputs
     val (inputs, outputs) = table.unzip
@@ -259,7 +167,7 @@ class QMCMinimizer extends Minimizer {
 
       val (essentialPrimeImplicants, nonessentialPrimeImplicants, uncoveredImplicants) =
         QMCMinimizer.getEssentialPrimeImplicants(
-          if (defaultToDc) primeImplicants.flatMap(_.expand(maxt)) else primeImplicants,
+          if (defaultToDc) primeImplicants.map(a => a.expand(maxt).reduce(_.mergeIfSimilar(_).getOrElse(a))) else primeImplicants,
           implicants
         )
